Use renamed file.
